+++ framework/src/Makefile
@@ -15,6 +15,7 @@ BS_LIB_SOURCES += fwk_interrupt.c
 BS_LIB_SOURCES += fwk_mm.c
 BS_LIB_SOURCES += fwk_module.c
 BS_LIB_SOURCES += fwk_slist.c
+BS_LIB_SOURCES += fwk_thread_delayed_resp.c
 ifeq ($(BUILD_HAS_MULTITHREADING),yes)
     BS_LIB_SOURCES += fwk_multi_thread.c
 else
